docs: Add fixes and suggestions for no-unsafe-negation
Created /fixes-and-suggestions/no-unsafe-negation.md under /user-guide,
and added fixes and suggestions using information from the rule's test source.

This file is a first draft to verify the content.
Once approved, it will be added to the main eslint repo,
and future PRs for this file will be accepted only from there.

There are also parts that could be removed,
and instead generated through the main eslint repo's Makefile.
Some of these parts are:

- front matter
- comment for non-native file
- Version section
- Resources section

Refs #899